+++ GameService.Domain/ValueObject/Position.cs
@@ -2,23 +2,23 @@ namespace GameService.Domain.ValueObject
 {
     public class Position
     {
-        public int X;
-        public int Y;
-        public int Z;
+        public decimal X;
+        public decimal Y;
+        public decimal Z;
 
-        public Position(int x, int y, int z)
+        public Position(decimal x, decimal y, decimal z)
         {
             X = x;
             Y = y;
             Z = z;
         }
 
-        public void Set(int x, int y, int z)
+        public void Set(decimal x, decimal y, decimal z)
         {
             (X,Y,Z) = (x,y,z);
         }
         
-        public (int, int, int) Get()
+        public (decimal, decimal, decimal) Get()
         {
             return (X,Y,Z);
         }
